feat: 검색 추가 및 정렬 전용 컴포넌트 분리
